OcxzvdEL - feat: structural changes on sensors from sensor library

+++ src/input/sensors/sensor_library/gy_ml8511/gy_ml8511.cpp
@@ -9,10 +9,10 @@ void gy_ml8511_init()
 float gy_ml8511_readUvIntensity()
 {
   int analog_value = analogRead(SENSORS_GY_ML8511_PIN_ANALOG);
-  float uv_voltage = (analog_value / GY_ML8511_ANALOG_INPUT_MAX) * GY_ML8511_VCC_VOLTAGE;  //Convert to voltage
+  float uv_voltage = ((float)analog_value / GY_ML8511_ANALOG_INPUT_MAX) * GY_ML8511_VCC_VOLTAGE;  //Convert to voltage
   
   // Convert voltage to intensity (UV intensity in mW/cm^2)
-  float calculated_intensity = map(uv_voltage, GY_ML8511_OUTPUT_VOLTAGE_MIN, GY_ML8511_OUTPUT_VOLTAGE_MAX, 
+  float calculated_intensity = (float)map(uv_voltage, GY_ML8511_OUTPUT_VOLTAGE_MIN, GY_ML8511_OUTPUT_VOLTAGE_MAX, 
                                    SENSORS_GYML8511_UV_MIN, SENSORS_GYML8511_UV_MAX);  // range from datasheet
 
   return calculated_intensity;
